pyproject.toml
[project]
name = "tasks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyright>=1.1.400",
]
